Improve OpenWRT installation

+++ openwrt/build-opk.sh
@@ -1,12 +1,19 @@
 #!/bin/bash -e
 
-./scripts/feeds update -a
-./scripts/feeds install -a
-./scripts/feeds install asterisk-chan-quectel
+PKG_NAME=asterisk-chan-quectel
 
+if [ -f .config ]; then
+    make package/${PKG_NAME}/clean
+    make package/${PKG_NAME}/compile
+    make package/index
+else
+    cat feeds-strskx.conf >> feeds.conf.default
+    ./scripts/feeds update -a
+    ./scripts/feeds install -a
+    ./scripts/feeds install ${PKG_NAME}
 
-cp diffconfig .config
-make defconfig
-make package/asterisk-chan-quectel/compile
-make package/index
-
+    cp diffconfig .config
+    make defconfig
+    make package/${PKG_NAME}/compile
+    make package/index
+fi
